move pad_contrast_matrix to _utils
